+++ SDK/AppCenterCrashes/Microsoft.AppCenter.Crashes.Shared/AndroidErrorDetails.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
+
 namespace Microsoft.AppCenter.Crashes
 {
     /// <summary>
@@ -8,9 +10,9 @@ namespace Microsoft.AppCenter.Crashes
     /// </summary>
     public class AndroidErrorDetails
     {
-        internal AndroidErrorDetails(object throwable, string threadName)
+        internal AndroidErrorDetails(string stackTrace, string threadName)
         {
-            Throwable = throwable;
+            StackTrace  = stackTrace;
             ThreadName = threadName;
         }
 
@@ -18,7 +20,14 @@ namespace Microsoft.AppCenter.Crashes
         /// Gets the throwable associated with the Java crash.
         /// </summary>
         /// <value>The throwable associated with the crash. <c>null</c> if the crash occured in .NET code.</value>
+        [Obsolete("This property is no longer set due to a security issue, use StackTrace instead.")]
         public object Throwable { get; }
+
+        /// <summary>
+        /// Gets the stack trace associated with the Java crash.
+        /// </summary>
+        /// <value>The stack trace associated with the crash. <c>null</c> if the crash occured in .NET code.</value>
+        public string StackTrace { get; }
 
         /// <summary>
         /// Gets the name of the thread that crashed.